feat: add auth util in cart service
auth util used to fetch the users email and profile id from the jwt token

+++ EShoppingZone/Cart-Service/src/main/java/com/eshopingzone/cartservice/model/CartItem.java
@@ -21,11 +21,11 @@ public class CartItem {
 
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private int cartItemId;
+	private Long cartItemId;
 	private double productPrice;
-	private int quantity;
+	private Integer quantity;
 	private double discount;
-	private int productId;
+	private Long productId;
 	
 	@ManyToOne
 	@JoinColumn(name = "cart_id")
